backend/requirements.txt
Flask
flask-restful
mysql-connector-python